requirements.txt
aiohttp==1.0.5
async-timeout==1.2.0
chardet==3.0.2
discord.py==0.16.7
multidict==2.1.4
websockets==3.3
yarl==0.9.2
